pyproject.toml
[project]
name = "researchclaw"
version = "0.5.0"
description = "ResearchClaw — Autonomous Research Pipeline. Turn any research idea into a paper."
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "rich>=13.0",
    "openai>=1.68.0",
    "azure-identity>=1.20.0",
]
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
researchclaw = "researchclaw.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["researchclaw", "sibyl", "arc"]

[tool.hatch.build.targets.wheel.force-include]
"researchclaw/templates/styles" = "researchclaw/templates/styles"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
